+++ src/components/memoBlock/memo.component.tsx
@@ -1,6 +1,5 @@
 import { useState, useEffect } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import cloneDeep from "lodash.clonedeep";
 import { AppDispatch } from "store";
 import { getImages } from "store/getImages";
 import { Props, Image, Level } from "components/interfaces";
@@ -9,6 +8,7 @@ import Blocks from "components/blocks";
 import RadioButtons from "components/radioButtons";
 import Title from "components/title";
 import { shuffleImageMemoryBlocks } from "helpers/shuffleImageMemoryBlocks";
+import handleMemoBlockClick from "helpers/memoGameLogic";
 
 interface MemoBlock {
   index: number;
@@ -23,11 +23,7 @@ function MemoBlockComponent(props: Props): JSX.Element {
   const dispatch: AppDispatch = useDispatch<AppDispatch>();
   const [shuffledMemoBlocks, setShuffledMemoBlocks] = useState<MemoBlocks>([]);
   const [selectedLevel, setSelectedLevel] = useState<Level>({ row: 4, col: 8 });
-  const [selectedBlock, setSelectedBlock] = useState<{
-    index: number;
-    image: Image;
-    flipped: boolean;
-  } | null>(null);
+  const [selectedBlock, setSelectedBlock] = useState<MemoBlock | null>(null);
   const [isAnimating, setAnimating] = useState(false);
 
   useEffect(() => {
@@ -50,68 +46,14 @@ function MemoBlockComponent(props: Props): JSX.Element {
   }
 
   function handleClick(memoBlock: MemoBlock) {
-    setShuffledMemoBlocks(
-      getSetWithMemoBlock(shuffledMemoBlocks, asFlipped(memoBlock))
-    );
-
-    if (noneSelection()) {
-      setSelectedBlock(memoBlock);
-    } else if (equalSelection()) {
-      // Pair selected, no need to keep tracking on it anymore.
-      setSelectedBlock(null);
-    } else {
-      flipAndResetSelection();
-    }
-
-    function getSetWithMemoBlock(memoBlocks: MemoBlocks, memoBlock: MemoBlock) {
-      memoBlocks = cloneDeep(memoBlocks);
-      memoBlocks.splice(memoBlock.index, 1, memoBlock);
-
-      return memoBlocks;
-    }
-
-    function asFlipped(memoBlock: MemoBlock) {
-      memoBlock = cloneDeep(memoBlock);
-      memoBlock.flipped = true;
-
-      return memoBlock;
-    }
-
-    function asNotFlipped(memoBlock: MemoBlock) {
-      memoBlock = asFlipped(memoBlock);
-      memoBlock.flipped = false;
-
-      return memoBlock;
-    }
-
-    function noneSelection() {
-      return selectedBlock === null || !selectedBlock;
-    }
-
-    function equalSelection() {
-      return selectedBlock && selectedBlock.image.id === memoBlock.image.id;
-    }
-
-    function flipAndResetSelection() {
-      setAnimating(true);
-
-      setTimeout(() => {
-        if (selectedBlock) {
-          let memoBlocksWithSelectionFlipped = getSetWithMemoBlock(
-            shuffledMemoBlocks,
-            asNotFlipped(memoBlock)
-          );
-          memoBlocksWithSelectionFlipped = getSetWithMemoBlock(
-            memoBlocksWithSelectionFlipped,
-            asNotFlipped(selectedBlock)
-          );
-
-          setShuffledMemoBlocks(memoBlocksWithSelectionFlipped);
-          setSelectedBlock(null);
-          setAnimating(false);
-        }
-      }, 1000);
-    }
+    handleMemoBlockClick({
+      memoBlock,
+      shuffledMemoBlocks,
+      selectedBlock,
+      setSelectedBlock,
+      setShuffledMemoBlocks,
+      setAnimating,
+    });
   }
 
   function handleSelectionLevels(level: string) {
